Move bodyPart trauma damage code into it's own file
