+++ UnityProject/Assets/Scripts/HealthV2/Living/Damage/BodyPartTraumaDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07471f83a315b0344b017a1ffea3332b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
